Time: 240 ms (15.12%), Space: 105.6 MB (44.19%) - LeetHub

+++ 2530-maximal-score-after-applying-k-operations/2530-maximal-score-after-applying-k-operations.java
@@ -1,16 +1,16 @@
 class Solution {
     public long maxKelements(int[] nums, int k) {
         long score = 0;
-        PriorityQueue<Long> heap = new PriorityQueue<>(Collections.reverseOrder());
+        PriorityQueue<Integer> heap = new PriorityQueue<>(Collections.reverseOrder());
         for(int i=0; i<nums.length; i++)
         {
-            heap.offer((long)nums[i]);
+            heap.offer(nums[i]);
         }
         for(int i=0; i<k; i++)
         {
-            long temp = heap.poll();
+            int temp = heap.poll();
             score+=temp;
-            heap.add((long)Math.ceil((temp*1.0)/3.0));
+            heap.add((int)Math.ceil((temp*1.0)/3.0));
         }
         return score;
     }
